mcp-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "infra-mcp-server"
version = "0.1.0"
description = "Infrastructure Self-Service MCP Server"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "uvicorn>=0.30.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0",
]

[project.scripts]
infra-mcp = "src.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]
